update my board config to the most complete version that works
this does NOT include any lists, as it seems like the current code's
attempt to splice two configs together is what is leading to the
corruption and causing bad board boots and/or breaking saves.

the most tangible version of this was seeing bad control characters in
the label for profile 1, if profile 1 was present both in the board
config and the user config --- instead of seeing the user config, I was
getting the board config's label w/corruption + board config's pin
mappings, and saves wouldn't work. committed board config works as
expected.

sadly one of the biggest benefits to doing the board configs was to put
profile configs in the build, which is still impossible until I figure
out what is going on with the code